esg-api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
fastapi==0.135.1
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multitasking==0.0.12
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.2
pluggy==1.6.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
yesg==2.1.1
yfinance==1.2.0
